% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apwnd_window_internal}
\alias{apwnd_window_internal}
\title{Compute average pairwise nucleotide diversity statistics for positions from a to b}
\usage{
apwnd_window_internal(x, a, b)
}
\arguments{
\item{x}{an 0,1,2,-1 integer matrix}

\item{a}{the lowest (base-1) index of the SNPs in the window}

\item{b}{the highest (base-1) index of the SNPs in the window}
}
\value{
This returns a list of vectors.  First vector is the average pairwise nucleotide difference
for each pair of individuals.  Second vector is the fraction of sites missing in the pair.
}
\description{
This is the internal Rcpp function that will be called once for each window.
}
