% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apwnd_windows.R
\name{apwnd_windows}
\alias{apwnd_windows}
\title{break markers up into non-overlapping windows and compute ave-pw-nuc-div}
\usage{
apwnd_windows(d012, width, positions = NULL)
}
\arguments{
\item{d012}{an 012 matrix.  Positions must be strictly increasing (so, no more than one
chromosome / scaffold)}

\item{width}{the desired window width in base pairs}

\item{positions}{can be used to pass the positions in if they aren't correctly parsed
in the 012 file}
}
\description{
more needed
}
