% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-selection.R
\name{comp_pair}
\alias{comp_pair}
\title{make a comparison between two groups that are in df}
\usage{
comp_pair(df, grp1, grp2)
}
\arguments{
\item{df}{the data frame you are picking things from.  This should be
like the data frame that you get back from group_012_cnts}

\item{grp1}{the name of the first group in the comparison}

\item{grp2}{the name of the second group in the comparison}
}
\value{
Sneds back a data frame sorted in descending order of
correct assignment prob.
}
\description{
boing
}
