% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-selection.R
\name{compile_assayable}
\alias{compile_assayable}
\title{return SNPs and say whether the are assayable or not, along with the MAFs}
\usage{
compile_assayable(w, minMAF = 0.03, flank = 20)
}
\arguments{
\item{w}{a matrix like produced by read_012}

\item{flank}{If a SNP is to be assayable, it needs to have no SNPs within flank
base pairs of itself.}

\item{MAF}{the minimum minor allele frequency to retain a SNP either for assaying
or for determining whether there is flanking variation.}
}
\value{
This returns a data frame with the columns:
\enumerate{
\item snp num_gene_copies
\item freq
\item scaffold
\item bp
\item length
\item leftpos
\item rightpos
\item leftpad
\item rightpad
\item assayable
\item on_right_edge
\item on_left_edge
}
}
\description{
This should be run on a set of data that has not been too highly filtered---you
want to know if there is much variation, even if it wasn't seen in everyone.
}

