% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-admixture-etc.R
\name{convert_012_to_bed}
\alias{convert_012_to_bed}
\title{create a PLINK bed file from and 012 matrix}
\usage{
convert_012_to_bed(W, chromo_override = FALSE, prefix = "plink")
}
\arguments{
\item{W}{the matrix that results from \code{\link{read_012}}.}

\item{chromo_override}{If TRUE, this sets the chromosome names to 1
so that ADMIXTURE will work.}

\item{prefix}{Prefix for the output files.}
}
\description{
This clearly does not preserve Ref and Alt alleles, etc., but can
be a quick way to get to a bed file.  It also allows the user to
override the chromosome names so that they play nicely with ADMIXTURE.
This function expects that plink is on the system PATH somewhere.
}
\details{
This function assumes everyone is unrelated and nobody has a phenotype
to be coded in plink.

It also currently is hardwired to deal with contig names the way they come
out of our pipeline and get catenated by \code{\link{read_012}}.  That is,
when it sees: \code{scaffold2|size4533513--4321992} it knows that the contig
comes before the -- and the position in bp comes after that.
}
