% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-selection.R
\name{correct_ass_prob}
\alias{correct_ass_prob}
\title{from vectors of snp allele freqs, compute the expected prob of correct assignment}
\usage{
correct_ass_prob(p0, p1)
}
\arguments{
\item{p0}{allele frequency in first population}

\item{p1}{allele frequency in second population}
}
\value{
a vector of the same length as p0 and p1
}
\description{
this is most useful for use with dplyr and mutate
}

