% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_freq_calcs.R
\name{geno_freq_calc_single}
\alias{geno_freq_calc_single}
\title{from on 012 file compute expected and observed genotype counts}
\usage{
geno_freq_calc_single(g012)
}
\arguments{
\item{g012}{an 012 matrix.  Missing data can be -1 or NA.}
}
\description{
...assuming HW equilbrium
}
