% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_freq_calcs.R
\name{geno_freq_calcs}
\alias{geno_freq_calcs}
\title{from on 012 file compute expected and observed genotype counts by population}
\usage{
geno_freq_calcs(g012, pops = NULL)
}
\arguments{
\item{g012}{an 012 matrix.  Missing data can be -1 or NA.}

\item{pops}{a data frame with one column "pop" and the other "sample" that tells
which populations individuals are in.  If "pop" entry is NA for an individual,
that individual is dropped silently. If NULL then everyone is assumed to be in the
same population.}
}
\description{
from on 012 file compute expected and observed genotype counts by population
}
