% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoscape_pca.R
\name{genoscape_pca}
\alias{genoscape_pca}
\title{do a PCA using SNPRelate and prepare output for rapid plotting}
\usage{
genoscape_pca(dat012, samples_to_remove = character(0), num_pcs = 6,
  sample_groups = NULL, plot_facet_free = FALSE)
}
\arguments{
\item{dat012}{an 012 matrix like that you get from read_012.
missing data should be denoted -1.}

\item{samples_to_remove}{the names of the samples that should be removed
before doing the PCA. This does not check to make sure that the samples
requested to be dropped actually exist in dat012.}

\item{num_pcs}{number of principal components to retain in output.  Default = 6}

\item{sample_groups}{If not NULL, this should be a tibble with a column "sample"
and another column "group", by which the points will be colored.}

\item{plot_facet_free}{if TRUE then the pca-pairs plot scales are free to encompass less area.}
}
\description{
This just does all the legwork to do a PCA using SNPRelate, and then
it does what it takes to plot every PC against every other using
facet_grid.  It doesn't make the plots, but it prepares the
data.
}
\details{
This function lets you say which individuals you want to drop which
makes it easy to iteratively remove outliers.
}
\examples{
Just while working on it, for testing:

dat012 <- read_012("../amke-popgen/data/rachael-amke-clean_indv175_pos110000", gz = TRUE)
sample_groups <- read_csv("../amke-popgen/data/amke_locations_from_mikki.csv")
samples_to_remove <- c("1833-14592", "2003-40521", "OCBPC-3", "OCBPC-4")
}
