% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoscapeRtools-package.R
\docType{package}
\name{genoscapeRtools}
\alias{genoscapeRtools}
\alias{genoscapeRtools-package}
\title{My package documentation
… bla bla …}
\description{
My package documentation
… bla bla …
}

