% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-admixture-etc.R
\name{ggplot_the_Qs}
\alias{ggplot_the_Qs}
\title{ggplot the Qs from multiple runs of ADMIXTURE}
\usage{
ggplot_the_Qs(Qs)
}
\arguments{
\item{Qs}{the data frame that comes from running \code{\link{slurp_admixture}}
on the Qs.}
}
\description{
This just plots everything onto one page and returns it as a ggplot object.
}

