% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-selection.R
\name{group_012_cnts}
\alias{group_012_cnts}
\title{count up number of observed alleles in each group}
\usage{
group_012_cnts(x012, glist)
}
\arguments{
\item{x012}{an 012 matrix like that returned by read_012}

\item{glist}{a named list of ID vectors. Each name is the name of the group
and each component is a vector of ID names that correspond to some of the individuals
in the rownames of x012.}
}
\value{
Returns a data frame with columns pos, n0, n1, group, ntot, and freq
}
\description{
boing.
}

