% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mat_cumul_cpp}
\alias{mat_cumul_cpp}
\title{return a matrix of cumumlative sums along the rows or colums}
\usage{
mat_cumul_cpp(x, dim)
}
\arguments{
\item{x}{a numeric matrix}

\item{dim}{1 for rows, 2 for columns}
}
\description{
return a matrix of cumumlative sums along the rows or colums
}
