% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-missingness.R
\name{miss_curves_locus}
\alias{miss_curves_locus}
\title{Order a genotype matrix by locus missingness at different numbers of individuals and plot results}
\usage{
miss_curves_locus(x012, reorder = TRUE, locus_thin_to = 5000,
  indv_nums = integer(0))
}
\arguments{
\item{x012}{a matrix of 012 genotypes with missing data coded as -1}

\item{reorder}{Logical.  if TRUE then the order of loci is re-done for every number of loci}

\item{locus_thin_to}{For speed with plotting it is possible to retain just  fraction of the loci,
once the cumulatives have been counted.  Default is 1000 of them.}

\item{indv_nums}{The number of individuals to explore retaining.  By default it just does them
in 10 roughly equal chunks.}
}
\description{
Order a genotype matrix by locus missingness at different numbers of individuals and plot results
}

