% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_fst.R
\name{pairwise_fst}
\alias{pairwise_fst}
\title{compute pairwise Fst using SNPRelate}
\usage{
pairwise_fst(dat012, sample_groups)
}
\arguments{
\item{dat012}{An 012 matrix.  Missing data should be -1.}

\item{sample_groups}{this should be a tibble with a column "sample"
and another column "group".  The Fst calculations are done pairwise between
the groups.  You probably don't want individuals in sample_groups that are
not in the rownames of dat012.  If group is NA for any individuals, they
get silently tossed out.}
}
\description{
This just wraps SNPRelate's function and makes it easy to compute
Fst for all pairwise combinations.  Returns a tidy data frame.
}
