% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_from_angsd_probs}
\alias{sample_from_angsd_probs}
\title{Sample a genotype (0,1,2) from probabilities for all the snps and individuals
in x.  This is a utility function.  It will get called by a higher level
function.}
\usage{
sample_from_angsd_probs(x, samples, snps, thresh)
}
\arguments{
\item{x}{a numeric vector that holds the posterior probs from an
ANGSD geno.gz file (from doGenos 32)}

\item{samples}{character vector of sample names}

\item{snps}{character vector of SNP names}

\item{thresh}{any snp in an individual must have one posterior greater than T in order to
have a non-missing value simulated for it.  For example, if T is 0.6 and the posteriors
are 0.33, 0.33, 0.33, this locus will just get a missing (-1).}
}
\description{
Sample a genotype (0,1,2) from probabilities for all the snps and individuals
in x.  This is a utility function.  It will get called by a higher level
function.
}
