% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_012.R
\name{scan_012}
\alias{scan_012}
\title{read a set of 012 files from vcftools into a matrix of genotypes with appropriate dimnames}
\usage{
scan_012(prefix, gz = FALSE, gzpos = FALSE, posfile_columns)
}
\arguments{
\item{prefix}{the (path) prefix on the files to be read.  The files will be \code{prefix.012},
\code{prefix.012.indv}, and \code{prefix.012.pos}}

\item{gz}{if this is set to TRUE, then the function expects that the 012 file ends in 012.gz
and is gzipped.  For this to work, you have to have a decent shell on your system with \code{gzcat}
for gunzipping the file to a stream.  If you are on a Mac, that should be fine.  If you are on
Windows, probably not.  If there are any problems, don't keep your file gzipped!}

\item{gzpos}{logical flag saying whether the pos file is gzipped.}

\item{posfile_columns}{The number of columns in the .pos file.  This seems to usually be 1 or 2.
If this is 1, then it is assumed that the locus names are unique.  If it is greater than 1, then
the first two columns are catenated together with "--" to form the position names.}
}
\value{
Returns a matrix with n-indiv rows and n-loci columns.  The rownames are the indv names and the
colnames are the pos's.
}
\description{
This one skips data.table::fread and just scans things directly.
Note that the individual names should be unique! So should the
locus names.  The .pos file should have two columns of names (at least things are hard-wired for
that at the moment).
}
