% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-admixture-etc.R
\name{slurp_admixture}
\alias{slurp_admixture}
\title{read in all the AMIXTURE files that are in the r_X_K_Y directories produced by run_admixture()}
\usage{
slurp_admixture(path = "admixture_runs", which_dat = "Q")
}
\arguments{
\item{path}{the directory that holds all the r_X_K_y directories.}

\item{which_dat}{By default, this gets the Q values.  If you want to get
the allele frequencies, set this to "P"}
}
\description{
Reads it into a tidy data frame
}
