% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore-missingness.R
\name{miss_curves_indv}
\alias{miss_curves_indv}
\title{Order a genotype matrix by individual missingness at different locus numbers and plot results}
\usage{
miss_curves_indv(x012, reorder = TRUE, loc_nums = integer(0))
}
\arguments{
\item{x012}{a matrix of 012 genotypes with missing data coded as -1}

\item{reorder}{Logical.  if TRUE then the order of individuals is re-done for every number of loci}

\item{loc_nums}{vector of numbers of loci to investigate.  If not given, then it just does 10 equally-spaced
values.}
}
\description{
Order a genotype matrix by individual missingness at different locus numbers and plot results
}

