% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-admixture-etc.R
\name{run_admixture}
\alias{run_admixture}
\title{run admixture Reps times at each of K values}
\usage{
run_admixture(bed, Reps = 2, Kvals = 2:5, path = ".",
  outdir = "admixture_runs")
}
\arguments{
\item{bed}{the bed file to use. (With the .bed extension.  Example = "plink.bed".)  The .bim file has to be with it
in the same directory as well as the .fam file.
 These three files will both get
copied to "input.bed" and "input.bim" in the "data" directory in the outdir}

\item{path}{the path that gets you to the directory you want to put all the results in.}

\item{outdir}{the single name of the output directory you want}
}
\description{
Just a wrapper to mclapply over all this.
}

